move some utils to tendermint client
